Commitment reminder emails (#8)
This PR adds automated commitment reminders. When the first commitment
is added to a given call on the day it's scheduled, two emails will be
scheduled to be sent:

1. 2 hours after the call was scheduled to start
1. A half a week after the call was scheduled

The commitment email summarizes the commitments added to the given call
and the success of previous commitments for each user in the group.