Prevent duplicate invites for existing memberships
Closes #18

This PR fixes an issue where group invites were being sent to existing
members of a group. Upon accepting the invite, an additional membership
record was being created for that user. This caused the group to appear
twice in the user's group list.

This commit adds validations around the creation of memberships and
invites. Users will no longer receive invites for groups they belong to
and it is no longer possible to create multiple membership records for
the same user.