Fix inconsistent date time display (#21)
Closes #16

This PR addresses two issues with the call times being displayed in a
few different places. Additionally, I refactored the DateFormatter to
DateTimeFormatter for a more accurate name.

Part of the issue was with the `strftime` calls being made in the
`DateTimeFormatter` using the wrong syntax.

Another issue was related to how calls were being created. The intention
was to use `Group#call_time` to set the `scheduled_on` attribute
of new calls, but the group attribute was never being used. This commit
fixes this issue as well.