Fix intermittent spec failures
This PR fixes an issue where tests would intermittently fail when the
local machines date differed from the UTC date. The failing tests' setup
phase was running in the default (UTC) timezone, but when the webdriver
executed the tests, the browser would set the timezone to the system's
timezone. This caused objects' timestamps to match tomorrows date
relative to the system's time instead of today's date.

The solution we arrived at was to stub out Time.zone to only return the
UTC timezone object before each test example. The only exception this is
in the ApplicationController test where we want to be able to test that
the application's timezone is indeed being set by the browser's time
zone.

Co-authored-by: Eric Hanko <eric.hanko1@gmail.com>