Fix timer countdown
This commit fixes a bug in the timer countdown introduced when time
zone functionality was added to the application.

Previously, the `data-timer-ends-at` attribute was manually being
labelled as a UTC time stamp by appending "UTC" to a string formatted
time stamp. Now that the application sets the time zone based on the
users' time zone, that time stamp was no longer a UTC time zone even
though had "UTC" appended.

The solution was to convert the time stamp added to the data attribute
to a UTC time stamp string using built in Rails methods.