Add basic styling to user profile page
This commit adds some nicer styling to the user profile page, including
adding the user's avatar to the page.

We also refactored the `UserAvatar` class to accept a size argument so a
larger-size image can be used where desired. These changes include some
clean up of the tests for the `UserAvatar` class.

Co-authored-by: Kurt G. <soad1789@gmail.com>