Update to Rails 6.1 (#63)
This work upgrades rails to 6.1

The only real code changes are in c472479. `form_with` used to generate forms with `data-remote` attribute set to `true`. This is now set to false by default. The helper no longer has a `remote` option. That option is renamed to `local` and defaults to `true`. I had to go through and change that on a few of the remote forms and add `local: false`

So a what used to look like:
```ruby
form_with model: @model, remote: true
```
Now looks like:
```ruby
form_with model: @model, local: false
```