Fix call loading logic
This commit fixes an issue where trying to edit an existing call would
result in a server error.

The main issue was that the `edit` action was not included in the list
of actions used to control which logic we used to load and authorize a
call. This resulted in no call being loaded at all.

Another issue was found where the wrong method was being used to check
the inclusion of the current user in the loaded record's list of
users. This resulted in potentially any user being able to load those
group or call records. This has also been fixed